+++ admin_ui/src/utils.ts
@@ -1,4 +1,9 @@
-import { type Schema, type OrderByConfig, getType } from "@/interfaces"
+import {
+    type Schema,
+    type OrderByConfig,
+    getType,
+    isNullable
+} from "@/interfaces"
 import router from "./router"
 import moment from "moment"
 
@@ -135,13 +140,25 @@ export function convertFormValue(params: {
 }): any {
     let { key, value, schema } = params
 
+    if (value == "null") {
+        return null
+    }
+
     const property = schema.properties[key]
 
-    if (value == "null") {
-        value = null
-    } else if (property.extra?.nullable && value == "") {
-        value = null
-    } else if (getType(property) == "array") {
+    const nullable = property.extra?.nullable
+
+    if (nullable == true && value == "") {
+        return null
+    }
+
+    // For Piccolo custom forms, there is no `extra` attribute, instead we
+    // have to look at the OpenAPI schema:
+    if (nullable == undefined && isNullable(property) && value == "") {
+        return null
+    }
+
+    if (getType(property) == "array") {
         value = JSON.parse(String(value))
     }