+++ piccolo_admin/example/forms/__init__.py
@@ -3,6 +3,7 @@ from .csv import FORM as CSV_FORM
 from .email import FORM as EMAIL_FORM
 from .image import FORM as IMAGE_FORM
 from .selections import FORM as SELECTION_FORM
+from .nullable import FORM as MEGA_FORM
 
 FORMS = [
     CALCULATOR_FORM,
@@ -10,4 +11,5 @@ FORMS = [
     EMAIL_FORM,
     IMAGE_FORM,
     SELECTION_FORM,
+    MEGA_FORM,
 ]